+++ servers/mcp-neo4j-data-modeling/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 ### Changed
+* Update README with link to data modeling demo repo and workflow image
 
 ### Added